pyproject.toml
[tool.poetry]
name = "picov3"
version = "0.1.0"
description = "Your personal raider.io bot."
authors = ["anress <anjress@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
peewee = "^3.17.1"
requests = "^2.31.0"
pillow = "^10.3.0"
unidecode = "1.3.8"
python-dateutil = "2.9.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
